refactor(api): import dependencies
