+++ api/scripts/create-users-accounts-for-contest.js
@@ -3,7 +3,7 @@ import bluebird from 'bluebird';
 import * as userToCreateRepository from '../lib/infrastructure/repositories/user-to-create-repository.js';
 import * as authenticationMethodRepository from '../lib/infrastructure/repositories/authentication-method-repository.js';
 import * as userService from '../lib/domain/services/user-service.js';
-import * as encryptionService from '../lib/domain/services/encryption-service.js';
+import * as encryptionService from '../src/shared/domain/services/encryption-service.js';
 import { disconnect } from '../db/knex-database-connection.js';
 import * as url from 'url';
