+++ api/src/authentication/domain/services/refresh-token-service.js
@@ -1,10 +1,10 @@
 import bluebird from 'bluebird';
 import { randomUUID } from 'crypto';
-import { tokenService } from './token-service.js';
+import { tokenService } from '../../../shared/domain/services/token-service.js';
 
-import { config } from '../../config.js';
-import { UnauthorizedError } from '../../application/http-errors.js';
-import { temporaryStorage } from '../../infrastructure/temporary-storage/index.js';
+import { config } from '../../../../lib/config.js';
+import { UnauthorizedError } from '../../../../lib/application/http-errors.js';
+import { temporaryStorage } from '../../../../lib/infrastructure/temporary-storage/index.js';
 
 const refreshTokenTemporaryStorage = temporaryStorage.withPrefix('refresh-tokens:');
 const userRefreshTokensTemporaryStorage = temporaryStorage.withPrefix('user-refresh-tokens:');
