+++ api/src/shared/domain/services/encryption-service.js
@@ -1,6 +1,6 @@
 import bcrypt from 'bcrypt';
-import { config } from '../../config.js';
-import { PasswordNotMatching } from '../errors.js';
+import { config } from '../../../../src/shared/config.js';
+import { PasswordNotMatching } from '../../../../lib/domain/errors.js';
 const { bcryptNumberOfSaltRounds } = config;
 const hashPassword = function (password) {
   return bcrypt.hash(password, bcryptNumberOfSaltRounds);
