+++ api/tests/authentication/unit/domain/services/refresh-token-service_test.js
@@ -1,12 +1,12 @@
-import { expect, sinon, catchErr } from '../../../test-helper.js';
-import { tokenService } from '../../../../lib/domain/services/token-service.js';
-import { config as settings } from '../../../../lib/config.js';
-import * as refreshTokenService from '../../../../lib/domain/services/refresh-token-service.js';
-import { UnauthorizedError } from '../../../../lib/application/http-errors.js';
+import { expect, sinon, catchErr } from '../../../../test-helper.js';
+import { tokenService } from '../../../../../src/shared/domain/services/token-service.js';
+import { config as settings } from '../../../../../src/shared/config.js';
+import * as refreshTokenService from '../../../../../src/authentication/domain/services/refresh-token-service.js';
+import { UnauthorizedError } from '../../../../../lib/application/http-errors.js';
 const refreshTokenTemporaryStorage = refreshTokenService.refreshTokenTemporaryStorage;
 const userRefreshTokensTemporaryStorage = refreshTokenService.userRefreshTokensTemporaryStorage;
 
-describe('Unit | Domain | Service | Refresh Token Service', function () {
+describe('Unit | Authentication | Domain | Services | Refresh Token Service', function () {
   describe('#createRefreshTokenFromUserId', function () {
     it('should create refresh access token with user id and source', async function () {
       // given
