+++ api/tests/shared/integration/infrastructure/repositories/admin-member-repository_test.js
@@ -1,13 +1,12 @@
-import { expect, databaseBuilder, knex, catchErr, sinon } from '../../../test-helper.js';
-import { PIX_ADMIN } from '../../../../lib/domain/constants.js';
+import { expect, databaseBuilder, knex, catchErr, sinon } from '../../../../test-helper.js';
+import { PIX_ADMIN } from '../../../../../lib/domain/constants.js';
+import * as adminMemberRepository from '../../../../../src/shared/infrastructure/repositories/admin-member-repository.js';
+import { AdminMember } from '../../../../../lib/domain/models/AdminMember.js';
+import { AdminMemberError } from '../../../../../lib/domain/errors.js';
 
 const { ROLES } = PIX_ADMIN;
 
-import * as adminMemberRepository from '../../../../lib/infrastructure/repositories/admin-member-repository.js';
-import { AdminMember } from '../../../../lib/domain/models/AdminMember.js';
-import { AdminMemberError } from '../../../../lib/domain/errors.js';
-
-describe('Integration | Infrastructure | Repository | adminMemberRepository', function () {
+describe('Integration | Shared | Infrastructure | Repositories | adminMemberRepository', function () {
   describe('#findAll', function () {
     it('should return all users with a pix admin role that are not disabled', async function () {
       // given
