+++ api/tests/test-helper.js
@@ -47,7 +47,7 @@ nock.disableNetConnect();
 
 import { buildLearningContent as learningContentBuilder } from './tooling/learning-content-builder/index.js';
 
-import * as tokenService from '../lib/domain/services/token-service.js';
+import * as tokenService from '../src/shared/domain/services/token-service.js';
 import { Membership } from '../lib/domain/models/index.js';
 
 const EMPTY_BLANK_AND_NULL = ['', '\t \n', null];
