+++ api/tests/unit/application/usecases/checkAdminMemberHasRolePixSupport_test.js
@@ -1,5 +1,5 @@
 import { expect, sinon } from '../../../test-helper.js';
-import { tokenService } from '../../../../lib/domain/services/token-service.js';
+import { tokenService } from '../../../../src/shared/domain/services/token-service.js';
 import * as useCase from '../../../../lib/application/usecases/checkAdminMemberHasRoleSupport.js';
 
 describe('Unit | Application | Use Case | checkAdminMemberHasRoleSupport', function () {
