+++ api/tests/unit/application/usecases/checkAdminMemberHasRoleSuperAdmin_test.js
@@ -1,5 +1,5 @@
 import { expect, sinon } from '../../../test-helper.js';
-import { tokenService } from '../../../../lib/domain/services/token-service.js';
+import { tokenService } from '../../../../src/shared/domain/services/token-service.js';
 import * as useCase from '../../../../lib/application/usecases/checkAdminMemberHasRoleSuperAdmin.js';
 
 describe('Unit | Application | Use Case | checkAdminMemberHasRoleSuperAdmin', function () {
