+++ api/tests/unit/domain/services/mail-service_test.js
@@ -2,7 +2,7 @@ import { sinon, expect } from '../../../test-helper.js';
 
 import * as mailService from '../../../../lib/domain/services/mail-service.js';
 import { mailer } from '../../../../src/shared/mail/infrastructure/services/mailer.js';
-import { tokenService } from '../../../../lib/domain/services/token-service.js';
+import { tokenService } from '../../../../src/shared/domain/services/token-service.js';
 import { config as settings } from '../../../../lib/config.js';
 import { getI18n } from '../../../tooling/i18n/i18n.js';
 import fr from '../../../../translations/fr.json' assert { type: 'json' };
