+++ api/tests/unit/domain/services/session-results-link-service_test.js
@@ -1,6 +1,6 @@
 import { expect, sinon } from '../../../test-helper.js';
 import * as sessionResultsLinkService from '../../../../lib/domain/services/session-results-link-service.js';
-import { tokenService } from '../../../../lib/domain/services/token-service.js';
+import { tokenService } from '../../../../src/shared/domain/services/token-service.js';
 import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 describe('Unit | Domain | Service | Session Results Link Service', function () {
